+++ include/PluginComponent.h
@@ -66,6 +66,7 @@ namespace PluginComponents {
         RESET_DETECTOR,
         QUICK_CONNECT,
         DEEP_SLEEP,
+        NTP,
         CONFIG,
         SERIAL2TCP,
         MAX
@@ -299,6 +300,8 @@ public:
                 return F("Quick Connect");
             case RTCMemoryId::SERIAL2TCP:
                 return F("Serial2TCP");
+            case RTCMemoryId::NTP:
+                return F("NTP");
             case RTCMemoryId::NONE:
             case RTCMemoryId::MAX:
                 break;