+++ src/kfc_firmware.cpp
@@ -420,7 +420,7 @@ void setup()
 #if DEBUG_DEEP_SLEEP
         __DBG_printf("wakeup=%u mode=%u", wakeup, deepSleepParams.getWakeupMode());
         if (deepSleepParams.getWakeupMode() == DeepSleep::WakeupMode::AUTO) {
-            Logger_notice(F("Wakeup from deep sleep, start-time=%u est.time=%u sleep-time=%.3f"), time(nullptr), deepSleepParams.getRealTime(), deepSleepParams.getTotalTime());
+            Logger_notice(F("Wakeup from deep sleep start-time=%u sleep-time=%.3f rtc-offset=%.6f"), time(nullptr), deepSleepParams.getTotalTime(), DeepSleep::_realTimeOffset / 1000000.0);
 
         }
 #endif