+++ src/plugins/ntp/ntp_plugin.cpp
@@ -116,6 +116,9 @@ void NTPPlugin::reconfigure(const String &source)
 void NTPPlugin::shutdown()
 {
     _checkTimer.remove();
+#if !RTC_SUPPORT
+    _rtcMemUpdate.remove();
+#endif
     settimeofday_cb(nullptr);
 }
 
@@ -151,6 +154,8 @@ void NTPPlugin::getStatus(Print &output)
 void NTPPlugin::execConfigTime()
 {
     settimeofday_cb(updateNtpCallback);
+
+
      // set refresh time to a minimum for the update. calling configTime() seems to ignore this value. once
      // the time is valid, it is reset to the regular NTP refresh interval
     _ntpRefreshTimeMillis = 15000;
@@ -165,7 +170,7 @@ void NTPPlugin::execConfigTime()
 void NTPPlugin::updateNtpCallback()
 {
     auto now = time(nullptr);
-    __LDBG_printf("new time=%u", (int)now);
+    __DBG_printf("new time=%u @%.3fs", (int)now, micros() / 1000000.0);
 
     if (IS_TIME_VALID(now)) {
         NTPPlugin::_ntpRefreshTimeMillis = Plugins::NTPClient::getConfig().getRefreshIntervalMillis();
@@ -175,6 +180,18 @@ void NTPPlugin::updateNtpCallback()
 #if RTC_SUPPORT
     // update RTC
     config.setRTC(now);
+#elif NTP_STORE_STATUS
+    NtpStatus ntp(now);
+    RTCMemoryManager::write(RTCMemoryManager::RTCMemoryId::NTP, &ntp, sizeof(ntp));
+#endif
+#if !RTC_SUPPORT
+    // emulate RTC using the RTC memory
+    // the time is restored during reboot
+    // this does not work when the RTC memory is lost and is very unprecise using deep sleep
+    RTCMemoryManager::setWriteTime(true);
+    _Timer(plugin._rtcMemUpdate).add(Event::milliseconds(1000), true, [](Event::CallbackTimerPtr) {
+        RTCMemoryManager::writeTime();
+    });
 #endif
 
 #if NTP_LOG_TIME_UPDATE